+++ src/types/core/dispatch.d.ts
@@ -2,7 +2,7 @@ import { Decoder, DecoderType } from '../_types';
 
 export type Values<T extends object> = T[keyof T];
 
-export function disjointUnion<O extends { [key: string]: Decoder<any> }>(
+export function taggedUnion<O extends { [key: string]: Decoder<any> }>(
     field: string,
     mapping: O,
 ): Decoder<Values<{ [key in keyof O]: DecoderType<O[key]> }>>;