+++ src/types/index.d.ts
@@ -20,7 +20,6 @@ export { constant, hardcoded, mixed, null_, undefined_, unknown } from './core/c
 export { date, iso8601 } from './core/date';
 export { describe } from './core/describe';
 export { dict, exact, inexact, mapping, object, pojo } from './core/object';
-export { disjointUnion } from './core/dispatch';
 export { either, oneOf } from './core/either';
 export { email, httpsUrl, nonEmptyString, regex, string, url } from './core/string';
 export { fail } from './core/fail';
@@ -29,4 +28,5 @@ export { integer, number, positiveInteger, positiveNumber } from './core/number'
 export { json, jsonObject, jsonArray } from './core/json';
 export { lazy } from './core/lazy';
 export { maybe, nullable, optional } from './core/optional';
+export { taggedUnion } from './core/dispatch';
 export { tuple } from './core/tuple';